Change seobs pkey and indexing
Hypertable on id made no sense since this column is not needed for any queries;
instead, use obsid. This also changes primary key since the hypertable column
must be part of pkey -> use composite pkey of obsid and seid.
In fact, the id column is not necessary, but it is kept here for now
to enable tracking a record back to LOTJU raw data if ever needed.